+++ packages/ui/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
-## [unreleased]
+## [3.3.0-alpha.5] - 2020-04-06
+
+No significant changes.
+
+## [3.3.0-alpha.4] - 2020-03-30
 
 ### Deprecated